Backend/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
Deprecated==1.2.14
Django==5.1.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
paypalrestsdk==1.13.3
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pyOpenSSL==24.2.1
pytz==2024.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vonage==3.17.1
vonage-jwt==1.1.3
vonage-utils==1.1.4
wrapt==1.16.0
gunicorn